# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Translator Studio API — mission control for the Subarashiki Hibi PT-BR fan translation
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Settings {
  baseUrl: string;
  hasApiKey: boolean;
  /** @nullable */
  model: string | null;
  models: string[];
  concurrency: number;
  batchSize: number;
  /** @nullable */
  lastDetectedAt: string | null;
  /** @nullable */
  lastDetectError: string | null;
}

export interface SettingsUpdate {
  baseUrl?: string;
  apiKey?: string;
  model?: string;
  /** @minimum 1 */
  concurrency?: number;
  /** @minimum 1 */
  batchSize?: number;
}

export type ScriptFileStatus = typeof ScriptFileStatus[keyof typeof ScriptFileStatus];


export const ScriptFileStatus = {
  pending: 'pending',
  queued: 'queued',
  translating: 'translating',
  done: 'done',
  error: 'error',
} as const;

export interface ScriptFile {
  filename: string;
  status: ScriptFileStatus;
  totalLines: number;
  translatedLines: number;
  skippedLines: number;
  /** @nullable */
  errorMessage: string | null;
  updatedAt: string;
}

export interface Stats {
  totalFiles: number;
  doneFiles: number;
  errorFiles: number;
  translatingFiles: number;
  totalLines: number;
  translatedLines: number;
}

export interface TranslateAllResult {
  queued: number;
}