# lib/api-zod/src/generated/types/scriptFile.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Translator Studio API — mission control for the Subarashiki Hibi PT-BR fan translation
 * OpenAPI spec version: 0.1.0
 */
import type { ScriptFileStatus } from './scriptFileStatus';

export interface ScriptFile {
  filename: string;
  status: ScriptFileStatus;
  totalLines: number;
  translatedLines: number;
  skippedLines: number;
  /** @nullable */
  errorMessage: string | null;
  updatedAt: Date;
}