# lib/api-zod/src/generated/types/settings.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Translator Studio API — mission control for the Subarashiki Hibi PT-BR fan translation
 * OpenAPI spec version: 0.1.0
 */

export interface Settings {
  baseUrl: string;
  hasApiKey: boolean;
  /** @nullable */
  model: string | null;
  models: string[];
  concurrency: number;
  batchSize: number;
  /** @nullable */
  lastDetectedAt: Date | null;
  /** @nullable */
  lastDetectError: string | null;
}