# lib/api-zod/src/generated/types/settingsUpdate.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Translator Studio API — mission control for the Subarashiki Hibi PT-BR fan translation
 * OpenAPI spec version: 0.1.0
 */

export interface SettingsUpdate {
  baseUrl?: string;
  apiKey?: string;
  model?: string;
  /** @minimum 1 */
  concurrency?: number;
  /** @minimum 1 */
  batchSize?: number;
}