# lib/api-zod/src/generated/types/stats.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Translator Studio API — mission control for the Subarashiki Hibi PT-BR fan translation
 * OpenAPI spec version: 0.1.0
 */

export interface Stats {
  totalFiles: number;
  doneFiles: number;
  errorFiles: number;
  translatingFiles: number;
  totalLines: number;
  translatedLines: number;
}