# lib/api-zod/src/generated/types/translateAllResult.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Translator Studio API — mission control for the Subarashiki Hibi PT-BR fan translation
 * OpenAPI spec version: 0.1.0
 */

export interface TranslateAllResult {
  queued: number;
}